+++ packages/sdk-coverage-tests/src/generate/command.js
@@ -20,16 +20,37 @@ async function generate({configPath, ...options}) {
 
     const tests = await prepareTests(config)
 
-    const emittedTests = emitTests(tests, {
+    if (tests.length <= 0) {
+      const message = `No test will be emitted. Please check "testPath", "emitSkipped", "emitOnly" config parameters`
+      console.log(chalk.yellow(message), '\n')
+      return
+    }
+
+    const {emittedTests, errors} = emitTests(tests, {
       makeSdk: config.initializeSdk,
       fileTemplate: config.testFrameworkTemplate,
     })
+
+    if (errors.length > 0) {
+      if (config.strict) {
+        const [{test, error}] = errors
+        console.log(chalk.red(`Error during emitting test "${test.name}":`))
+        throw error
+      }
+      errors.forEach(({test, error}) => {
+        console.log(chalk.red(`Error during emitting test "${test.name}":`))
+        console.log(chalk.grey(error.stack))
+      })
+    }
+
     await createTestFiles(emittedTests, config)
     await createTestMetaData(emittedTests, config)
 
     const skippedTestsCount = emittedTests.reduce((count, t) => count + (t.skip ? 1 : 0), 0)
 
-    console.log(`Created ${emittedTests.length} (${skippedTestsCount} skipped) test files`)
+    console.log(chalk.green(`${chalk.bold(`${emittedTests.length}`.padEnd(3))} test(s) generated`))
+    console.log(chalk.cyan(`${chalk.bold(`${skippedTestsCount}`.padEnd(3))} test(s) skipped`))
+    console.log(chalk.red(`${chalk.bold(`${errors.length}`.padEnd(3))} error(s) occurred`))
   } catch (err) {
     console.log(chalk.red(err.stack))
     process.exit(1)